Move assertions to separate file
