+++ EssentialApp/EssentialAppTests/Helpers/ImageCommentCell+TestHelpers.swift
@@ -1,0 +1,24 @@
+//
+//  ImageCommentCell.swift
+//  EssentialAppTests
+//
+//  Created by Cronay on 24.12.20.
+//  Copyright © 2020 Essential Developer. All rights reserved.
+//
+
+import UIKit
+import EssentialFeediOS
+
+extension ImageCommentCell {
+	var usernameText: String? {
+		usernameLabel.text
+	}
+
+	var messageText: String? {
+		messageLabel.text
+	}
+
+	var dateText: String? {
+		dateLabel.text
+	}
+}
